Change project name to agda-ocaml
This commit changes the name of the project to agda-ocaml as well as
reverts some changes made to the project to make it easier to merge
into `agda` -- but removed the definition of the main executable in
this project.

It also reverts the LICENSE- and README files to before making the attempt
to integrate it.